Main Menu Skeleton is functional
Main menu contains three buttons (play, settings, and quit). Clicking "play" will begin the scene currently labeled "tucker_2" which is the bedroom hub. "quit" should quit the application and send a message to the debug log, however in Unity I don't think this closes the tester. The settings menu exists but there is an issue assigning "SettingsMenu" variable with the "SettingsMenu" gameobject. So the settings button is inoperable.