Bug Fix
Fixed Bug where ChestMessage button had to be pressed twice for it to close. So, when the Player exited the Box Collider, the chest would close and be destroyed, but when the button was pressed once and the Chest destroyed, the Player technically left the collider, which would try to close the UI but would instead open it back up.