HeavyWeight Item
Added new HeavyWeigth Item that boosts Player Defense but drops Player MoveSpeed. Came to predicament that happens when players have multiple items active that increase and decrease move speed. Made it to where SuperHuman and HeavyWeight will check the cock if each other are active and get the player's actual move speed from the other if so.